+++ src/Graft.VS2026Extension/ToolWindows/StackExplorerControl.xaml.cs
@@ -1,0 +1,14 @@
+using System.Windows.Controls;
+using Graft.VS2026Extension.Graft;
+
+namespace Graft.VS2026Extension.ToolWindows
+{
+    public partial class StackExplorerControl : UserControl
+    {
+        internal StackExplorerControl(GraftService? service)
+        {
+            InitializeComponent();
+            DataContext = new StackExplorerViewModel(service);
+        }
+    }
+}